requirements.txt
google-generativeai
pygame
python-dotenv